feature: 視界範囲3.5に

+++ Assets/Scripts/SO/SO_DifficultySettings.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Sirenix.OdinInspector;
 using UnityEngine;
 
 namespace SO
@@ -37,7 +36,7 @@ namespace SO
         [SerializeField, Header("難易度設定\nE, N, H, N")]
         private Difficulty[] difficulty;
 
-        public int VisibilityRange => difficulty[Difficulty.Type.ToInt()].VisibilityRange;
+        public float VisibilityRange => difficulty[Difficulty.Type.ToInt()].VisibilityRange;
         public float StaminaRecover => difficulty[Difficulty.Type.ToInt()].StaminaRecover;
         public bool IsItemRandom => difficulty[Difficulty.Type.ToInt()].IsItemRandom;
         public bool IsExtraShoggothArrangeDifficult => difficulty[Difficulty.Type.ToInt()].IsExtraShoggothArrangeDifficult;
@@ -46,7 +45,7 @@ namespace SO
     [Serializable]
     public class Difficulty
     {
-        [Header("視界範囲")] public int VisibilityRange;
+        [Header("視界範囲")] public float VisibilityRange;
         [Header("スタミナ回復速度")] public float StaminaRecover;
         [Header("アイテム配置ランダム化")] public bool IsItemRandom;
         [Header("エクストラショゴスの配置難化")] public bool IsExtraShoggothArrangeDifficult;
